+++ eras/babbage/impl/test/Test/Cardano/Ledger/Babbage/BinarySpec.hs
@@ -7,13 +7,11 @@ module Test.Cardano.Ledger.Babbage.BinarySpec (spec) where
 
 import Cardano.Ledger.Alonzo.TxWits (Redeemers, TxDats)
 import Cardano.Ledger.Babbage
-import Data.Default (def)
 import Test.Cardano.Ledger.Alonzo.Binary.RoundTrip (roundTripAlonzoCommonSpec)
 import Test.Cardano.Ledger.Babbage.Arbitrary ()
 import Test.Cardano.Ledger.Babbage.Binary.Annotator ()
 import Test.Cardano.Ledger.Babbage.TreeDiff ()
 import Test.Cardano.Ledger.Common
-import Test.Cardano.Ledger.Core.Binary (specUpgrade)
 import Test.Cardano.Ledger.Core.Binary as Binary (
   decoderEquivalenceCoreEraTypesSpec,
   decoderEquivalenceEraSpec,
@@ -23,7 +21,6 @@ import Test.Cardano.Ledger.Core.Binary.RoundTrip (RuleListEra (..))
 
 spec :: Spec
 spec = do
-  specUpgrade @BabbageEra def
   describe "RoundTrip" $ do
     roundTripAlonzoCommonSpec @BabbageEra
   describe "DecCBOR instances equivalence" $ do