+++ eras/conway/impl/testlib/Test/Cardano/Ledger/Conway/BinarySpec.hs
@@ -15,7 +15,6 @@ import Cardano.Ledger.Conway.Genesis
 import Cardano.Ledger.Conway.Governance
 import Cardano.Ledger.Core
 import Cardano.Ledger.Shelley.LedgerState (StashedAVVMAddresses)
-import Data.Default (def)
 import Data.Proxy
 import Data.Typeable (typeRep)
 import Test.Cardano.Ledger.Binary (decoderEquivalenceExpectation)
@@ -24,17 +23,14 @@ import Test.Cardano.Ledger.Common
 import Test.Cardano.Ledger.Conway.Arbitrary ()
 import Test.Cardano.Ledger.Conway.Binary.Annotator ()
 import Test.Cardano.Ledger.Conway.Binary.RoundTrip (roundTripConwayCommonSpec)
-import Test.Cardano.Ledger.Conway.Era (BabbageEraTest)
 import Test.Cardano.Ledger.Conway.ImpTest (ConwayEraImp)
 import Test.Cardano.Ledger.Conway.TreeDiff ()
-import Test.Cardano.Ledger.Core.Binary (specUpgrade)
 import Test.Cardano.Ledger.Core.Binary as Binary (decoderEquivalenceCoreEraTypesSpec, txSizeSpec)
 import Test.Cardano.Ledger.Core.Binary.RoundTrip (RuleListEra, roundTripEraSpec)
 
 spec ::
   forall era.
-  ( BabbageEraTest (PreviousEra era)
-  , ConwayEraImp era
+  ( ConwayEraImp era
   , DecCBOR (TxAuxData era)
   , DecCBOR (TxWits era)
   , DecCBOR (TxBody era)
@@ -46,7 +42,6 @@ spec ::
   ) =>
   Spec
 spec = do
-  specUpgrade @era def
   describe "RoundTrip" $ do
     roundTripCborSpec @GovActionId
     roundTripCborSpec @(GovPurposeId 'PParamUpdatePurpose)