+++ eras/dijkstra/src/Cardano/Ledger/Dijkstra/TxWits.hs
@@ -9,10 +9,8 @@ import Cardano.Ledger.Alonzo.TxWits (
   datsAlonzoTxWitsL,
   rdmrsAlonzoTxWitsL,
   scriptAlonzoTxWitsL,
-  upgradeRedeemers,
-  upgradeTxDats,
  )
-import Cardano.Ledger.Conway.Core (AlonzoEraTxWits (..), EraScript (..), EraTxWits (..))
+import Cardano.Ledger.Conway.Core (AlonzoEraTxWits (..), EraTxWits (..))
 import Cardano.Ledger.Conway.TxWits (AlonzoTxWits (..))
 import Cardano.Ledger.Dijkstra.Era (DijkstraEra)
 import Cardano.Ledger.Dijkstra.Scripts ()
@@ -30,15 +28,6 @@ instance EraTxWits DijkstraEra where
 
   scriptTxWitsL = scriptAlonzoTxWitsL
   {-# INLINE scriptTxWitsL #-}
-
-  upgradeTxWits atw =
-    AlonzoTxWits
-      { txwitsVKey = txwitsVKey atw
-      , txwitsBoot = txwitsBoot atw
-      , txscripts = upgradeScript <$> txscripts atw
-      , txdats = upgradeTxDats (txdats atw)
-      , txrdmrs = upgradeRedeemers (txrdmrs atw)
-      }
 
 instance AlonzoEraTxWits DijkstraEra where
   datsTxWitsL = datsAlonzoTxWitsL